Updated comments and identifies authors of each class

+++ src/main/java/com/juaracoding/smartpro_rest_api/model/ProcurementRequest.java
@@ -31,7 +31,7 @@ public class ProcurementRequest {
     private LocalDate expectedDate;
 
     @Column(name = "Status", nullable = false)
-    private Integer status;
+    private Short status;
 
     @Column(name = "CreatedBy", nullable = false, updatable = false)
     private Long createdBy;
@@ -101,11 +101,11 @@ public class ProcurementRequest {
         this.expectedDate = expectedDate;
     }
 
-    public Integer getStatus() {
+    public Short getStatus() {
         return status;
     }
 
-    public void setStatus(Integer status) {
+    public void setStatus(Short status) {
         this.status = status;
     }
